requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
pyserial>=3.5





